requirements.txt
pandas
threaded
speedtest-cli
workalendar
pylint